+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 0.8.178 (2021-12-07 / f93ce7b)
 
 ## Fixed